Module neon_runtime::call

Facilities for working with v8::FunctionCallbackInfo and getting the current v8::Isolate.

Functions

callee⚠
Mutates the out argument provided to refer to the v8::Local handle value of the currently executing function.

current_isolate⚠
Gets the current v8::Isolate.

data⚠
Mutates the out argument provided to refer to the v8::Local handle value of the v8::FunctionCallbackInfo Data.

get⚠
Mutates the out argument provided to refer to the v8::Local handle value of the ith argument passed to the function.

get_isolate⚠
Gets the isolate of the function call.

is_construct⚠
Indicates if the function call was invoked as a constructor.

len⚠
Gets the number of arguments passed to the function.

set_return⚠
Sets the return value of the function call.

this⚠
Mutates the out argument provided to refer to the v8::Local handle value of the object the function is bound to.