Module neon_runtime::error

Facilities for creating and throwing JS errors.

Functions

new_error⚠
Mutates the out argument provided to refer to a newly created Error object.

new_range_error⚠
Mutates the out argument provided to refer to a newly created RangeError object.

new_reference_error⚠
Mutates the out argument provided to refer to a newly created ReferenceError object.

new_syntax_error⚠
Mutates the out argument provided to refer to a newly created SyntaxError object.

new_type_error⚠
Mutates the out argument provided to refer to a newly created TypeError object.

throw⚠
Throws an Error object in the current context.

throw_error_from_cstring⚠
Throws an Error object in the current context.

throw_range_error_from_cstring⚠
Throws a RangeError object in the current context.

throw_reference_error_from_cstring⚠
Throws a ReferenceError object in the current context.

throw_syntax_error_from_cstring⚠
Throws a SyntaxError object in the current context.

throw_type_error_from_cstring⚠
Throws a TypeError object in the current context.