Module neon_runtime::nan

Modules

array
Facilities for working with v8::Arrays.

arraybuffer
Facilities for working with v8::ArrayBuffers.

buffer
Facilities for working with node::Buffers.

call
Facilities for working with v8::FunctionCallbackInfo and getting the current v8::Isolate.

class

convert
Helper functions for converting v8::Local values.

error
Facilities for creating and throwing JS errors.

fun
Facilities for working with v8::Functions.

handler
Facilities for running a callback in the libuv main thread.

mem
A helper function for comparing v8::Local handles.

module
A helper function for initializing a module.

object
Facilities for working with v8::Objects.

primitive
Facilities for working with primitive values.

raw
Fundamental definitions for mapping to the V8 memory space.

scope
Facilities for working with v8::HandleScopes and v8::EscapableHandleScopes.

string
Facilities for working with v8::Strings.

tag
Facilities for identifying the type of a v8::Local handle.

task
Facilities for running background tasks in the libuv thread pool.