Module neon_runtime::nan::buffer

Facilities for working with node::Buffers.

Functions

data⚠
Mutates the base_out and size_out arguments to access the data of a node::Buffer object.

new⚠
Mutates the out argument provided to refer to a newly created and zero-filled node::Buffer object. Returns false if the value couldn't be created.

uninitialized⚠
Mutates the out argument provided to refer to a newly created node::Buffer object. Returns false if the value couldn't be created.