Module neon_runtime::nan::object

Facilities for working with v8::Objects.

Functions

get⚠
Mutates the out argument provided to refer to the v8::Local value at the key provided. Returns false if the result couldn't be retrieved.

get_index⚠
Mutates the out argument provided to refer to the v8::Local value at the index provided of the v8::Object. Returns false if the result couldn't be retrieved.

get_isolate⚠
Gets the v8::Isolate of a v8::Object.

get_own_property_names⚠
Mutates the out argument provided to refer to a newly created v8::Array containing the names of the v8::Object's own property names. Returns false if the result is empty.

get_string⚠
Mutates the out argument provided to refer to the v8::Local value of the v8::String's underlying content. Returns false if the value couldn't be retrieved.

new⚠
Mutates the out argument provided to refer to a newly created v8::Object.

set⚠
Sets the key value of a v8::Object at the key provided. Also mutates the out argument provided to refer to a v8::Local boolean value, true if the set was successful.

set_index⚠
Sets the key value of a v8::Object at the index provided. Also mutates the out argument provided to refer to a v8::Local boolean value, true if the set was successful.

set_string⚠
Sets the underlying content of a v8::String object. Also mutates the out argument provided to refer to a v8::Local boolean value, true if the set was successful.