Module neon_runtime::nan::primitive

Facilities for working with primitive values.

Functions

boolean⚠
Mutates the out argument provided to refer to the v8::Boolean object.

boolean_value⚠
Gets the underlying value of a v8::Boolean object.

integer⚠
Mutates the out argument provided to refer to a newly created v8::Integer object.

integer_value⚠
Gets the underlying value of a v8::Integer object.

is_i32⚠
Indicates if the value is a 32-bit signed integer.

is_u32⚠
Indicates if the value is a 32-bit unsigned integer.

null⚠
Mutates the out argument provided to refer to the v8::Null object.

number⚠
Mutates the out argument provided to refer to a newly created v8::Number object.

number_value⚠
Gets the underlying value of a v8::Number object.

undefined⚠
Mutates the out argument provided to refer to the v8::Undefined object.