Function neon_runtime::nan::try_catch::with

pub unsafe extern "C" fn with(
    glue: extern "C" fn(*mut c_void, *mut c_void, *mut Local, *mut *mut c_void) -> TryCatchControl, 
    rust_thunk: *mut c_void, 
    cx: *mut c_void, 
    result: *mut Local, 
    unwind_value: *mut *mut c_void
) -> TryCatchControl

Wraps a computation with an RAII-allocated Nan::TryCatch. Invokes a Rust closure with a TryCatch live on the stack. The result value can be assumed to be initialized if and only if this function does not return TryCatchControl::Panicked. The unwind_value value can be assumed to be initialized if and only if this function returns TryCatchControl::Panicked.