Function neon_runtime::napi::async_work::schedule

pub unsafe fn schedule<I, O, D>(
    env: Env,
    input: I,
    execute: fn(input: I) -> O,
    complete: fn(env: Env, output: Result<O>, data: D),
    data: D
)where
    I: Send + 'static,
    O: Send + 'static,
    D: Send + 'static,

Schedule work to execute on the libuv thread pool

Safety

env must be a valid napi_env for the current thread

The thread::Result::Err must only be used for resuming unwind if execute is not unwind safe