Module neon_runtime::napi::object

Functions

get⚠

Mutates out to refer to the value of the property of object named by the key value. Returns false if the value couldn’t be retrieved.

get_index⚠

Mutate the out argument to refer to the value at index in the given object. Returns false if the value couldn’t be retrieved.

get_own_property_names⚠

Mutates the out argument to refer to a napi_value containing the own property names of the object as a JavaScript Array.

get_string⚠

Mutate the out argument to refer to the value at a named key in the given object. Returns false if the value couldn’t be retrieved.

new⚠

Mutates the out argument to refer to a napi_value containing a newly created JavaScript Object.

set⚠

Sets the property value of an napi_value object, named by another value key. Returns true if the set succeeded.

set_index⚠

Sets the key value of a napi_value at the index provided. Returns true if the set succeeded.

set_string⚠

Sets the key value of a napi_value at a named key. Returns true if the set succeeded.