Module neon_runtime::napi::primitive

Functions

boolean⚠

Mutates the out argument provided to refer to one of the global true or false objects.

boolean_value⚠

Get the boolean value out of a Local object. If the Local object does not contain a boolean, this function panics.

null⚠

Mutates the out argument provided to refer to the global null object.

number⚠

Mutates the out argument provided to refer to a newly created Local containing a JavaScript number.

number_value⚠

Gets the underlying value of an Local object containing a JavaScript number. Panics if the given Local is not a number.

undefined⚠

Mutates the out argument provided to refer to the global undefined object.