Module neon_runtime::scope

Facilities for working with v8::HandleScopes and v8::EscapableHandleScopes.

Functions

alignment⚠
Gets the alignment requirement of a v8::HandleScope.

chained⚠
Creates a v8::EscapableHandleScope and calls the callback provided with the argument signature (out, parent_scope, &v8_scope, closure).

enter⚠
Instantiates a new v8::HandleScope.

escapable_alignment⚠
Gets the alignment requirement of a v8::EscapableHandleScope.

escapable_size⚠
Gets the size of a v8::EscapableHandleScope.

escape⚠
Mutates the out argument provided to refer to the newly escaped v8::Local value.

exit⚠
Destructs a v8::HandleScope.

nested⚠
Creates a v8::HandleScope and calls the callback provided with the argument signature (out, realm, closure).

size⚠
Gets the size of a v8::HandleScope.