Module neon_runtime::napi

Modules

array

Facilities for working with Array napi_values.

arraybuffer

async_work

Rust wrappers for Node-API simple asynchronous operations

buffer

call

convert

date

error

external

fun

Facilities for working with JS functions.

lifecycle

Environment life cycle APIs

mem

no_panic

Utilities that will not panic for use in contexts where unwinding would be undefined behavior.

object

primitive

promise

JavaScript Promise and Deferred handle

raw

reference

scope

string

tag

tsfn

Idiomatic Rust wrappers for N-API threadsafe functions

typedarray

Enums

TypedArrayType

Functions

setup⚠

Loads N-API symbols from host process. Must be called at least once before using any functions in neon-runtime or they will panic. Safety: env must be a valid napi_env for the current thread

Type Definitions

Deferred